backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "chromadb==1.0.13",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "distro==1.9.0",
    "durationpy==0.10",
    "filelock==3.18.0",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "flask-httpauth==4.8.0",
    "flatbuffers==25.2.10",
    "fsspec==2025.5.1",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.73.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.33.1",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "kubernetes==33.1.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numpy==2.3.1",
    "oauthlib==3.3.1",
    "onnxruntime==1.22.0",
    "openai>=2.14.0",
    "opentelemetry-api==1.34.1",
    "opentelemetry-exporter-otlp-proto-common==1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
    "opentelemetry-proto==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-semantic-conventions==0.55b1",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.0",
    "pillow==11.2.1",
    "posthog==6.0.0",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pypdf2==3.0.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rich==14.0.0",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.16.0",
    "sentence-transformers==4.1.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.2",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "transformers==4.53.0",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "waitress==3.0.2",
    "watchfiles==1.1.0",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "zipp==3.23.0",
]
